requirements.txt
Flask==3.1.0
gunicorn==23.0.0
openai==1.72.0
playwright==1.51.0
python-dotenv==1.1.0
requests==2.32.3

